fix golangci-lint errors: misspellings, ineffassign, unconvert, deprecated APIs
- Fixed 4 misspellings: determins→determines, behaviour→behavior, determinisitc→deterministic
- Added BIP39 wordlist exclusion to prevent 43 false positive misspell errors
- Fixed 4 ineffassign errors by adding missing error checks
- Fixed 3 unconvert errors by removing unnecessary type conversions
- Replaced deprecated runtime.GOROOT() with exec.Command("go", "env", "GOROOT")
- Removed unused version variable in skycoin-web
- Fixed goimports formatting in explorer.go and middleware_test.go
- Fixed tab character in middleware_test.go

Reduced from 252 to 197 issues (22% reduction)

+++ src/api/network.go
@@ -216,7 +216,7 @@ func disconnectHandler(gateway Gatewayer) http.HandlerFunc {
 			return
 		}
 
-		if err := gateway.DisconnectByGnetID(uint64(id)); err != nil {
+		if err := gateway.DisconnectByGnetID(id); err != nil {
 			switch err {
 			case daemon.ErrConnectionNotExist:
 				wh.Error404(w, "")
